Downgrade VMA to 2.3.0
VMA is broken on drivers that don't support Vulkan 1.3, even though it is set to use 1.1. No matter what, it tries to dynamically link to `vkGetDeviceMemoryBufferRequirements` and fails miserably on devices with no Vulkan 1.3 support. Generally this can be worked around by updating the GPU driver, but that's not always an option for older GPUs.