Fix incorrect drone data flight point interpolation
The returned flight point was obtained by interpolating the next point in time with the one after it (2 points ahead), leading to a negative lerp time difference being used and causing "jumpiness" when rendering altitude.
This was caused by `lower_bound` returning the first element that is (1) greater than or (2) equal to the requested value. We always fall into situation (1) because of using floats, meaning the next point was returned instead of the current point.
The issue has now been fixed by using `upper_bound` to get the next point, and then subtracting one to get the current point.